pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plaude-thunes"
version = "0.1.0"
description = "Reusable Python SDK for integrating with the Thunes payment API"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["thunes", "payments", "sdk", "django"]

dependencies = [
    "requests>=2.28.0",
    "djangorestframework>=3.14.0",
]

[project.optional-dependencies]
django = [
    "django>=4.0",
]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "responses>=0.23",
    "factory-boy>=3.2",
]

[tool.setuptools]
packages = [
    "plaude_thunes",
    "plaude_thunes.api",
    "plaude_thunes.api.views",
    "plaude_thunes.api.serializers",
    "plaude_thunes.clients",
    "plaude_thunes.security",
    "plaude_thunes.services",
    "plaude_thunes.utils",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
